+++ pdtable/test/io/parsers/test_block_parsers.py
@@ -1,5 +1,6 @@
 import numpy as np
 import pandas as pd
+import pytest
 import datetime as dt
 from textwrap import dedent
 
@@ -151,6 +152,53 @@ def test_make_table__no_trailing_sep():
     t = make_table(cells).df
     assert t.column[0] == "bar"
     assert t.dash[0] == 10
+
+
+def test_make_table__empty():
+    cells = [
+        ["**an_empty_table"],
+        ["all"]
+    ]
+    t = make_table(cells).df
+    assert t.shape == (0, 0)
+
+
+def test_make_table__empty_transposed():
+    cells = [
+        ["**an_empty_table*"],
+        ["all"]
+    ]
+    t = make_table(cells).df
+    assert t.shape == (0, 0)
+
+
+def test_make_table__no_units_raises():
+    cells = [
+        ["**an_empty_table"],
+        ["all"],
+        ["col_a", "col_b"]
+    ]
+    with pytest.raises(ValueError):
+        make_table(cells)
+
+
+def test_make_table__empty_values():
+    lines = [
+        ["**foo"],
+        ["all"],
+        ["place", "distance"],
+        ["text", "km"],
+    ]
+
+    t = make_table(lines)
+
+    assert t.name == "foo"
+    assert set(t.metadata.destinations) == {"all"}
+    assert t.column_names == ["place", "distance"]
+    assert t.units == ["text", "km"]
+
+    df = pd.DataFrame({"place": [], "distance": []})
+    pd.testing.assert_frame_equal(t.df, df)
 
 
 def test_parse_blocks():
@@ -453,9 +501,9 @@ def test_parse_blocks__block_types():
     ]
     # fmt: on
 
+    parse_result = list(parse_blocks(cell_rows, to="cellgrid"))
     seen = {}
-    for ty, block in parse_blocks(cell_rows, to="cellgrid"):
-        #  print(f"\n-oOo- {ty} {block}")
+    for ty, block in parse_result:
         if seen.get(ty) is None:
             seen[ty] = []
         seen[ty].append(block)